Change to Strict Number Type System
We were using generics to set the return type of an executable. This forced me
to rely on MiscUtil for executing operators, which only allows for a limited
subset of operations. I changed my builders to let users specify the return
type as either decimal, float, or double. Nodes also resolve explicitly to
one of those three types. This allows for a much wider range of possible
operations, built directly on .NET's native numeric types.